% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{adaptive_steps_dt}
\alias{adaptive_steps_dt}
\title{adaptive_steps_dt}
\usage{
adaptive_steps_dt(
  x2_vals,
  m,
  grid_mode = "auto",
  k_auto_exact_threshold = 1000,
  interval_mode = "divide"
)
}
\description{
adaptive_steps_dt
}
\details{
adaptive_steps_dt
}
