% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{bereken_matches_en_stats}
\alias{bereken_matches_en_stats}
\title{bereken_matches_en_stats}
\usage{
bereken_matches_en_stats(
  dt1,
  dt2,
  kol_a,
  kol_x1,
  kol_x2,
  m = 0.01,
  grid_mode,
  k_auto_exact_threshold = 1000,
  interval_mode
)
}
\description{
bereken_matches_en_stats
}
\details{
bereken_matches_en_stats
}
