% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{bewaar_plots_als_pdf}
\alias{bewaar_plots_als_pdf}
\title{Bewaar plots als pdf}
\usage{
bewaar_plots_als_pdf(
  plotlijst,
  indices,
  bestandsnaam = "plots.pdf",
  width = 8,
  height = 6
)
}
\description{
Bewaar plots als pdf
}
