% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{concat_columns}
\alias{concat_columns}
\title{Voeg samengestelde kolom toe door samenvoegen van bestaande kolommen}
\usage{
concat_columns(dt, keystring, new_col, sep = ".", verbose = FALSE)
}
\arguments{
\item{dt}{data.table waarin de nieuwe kolom toegevoegd wordt (wordt by-reference aangepast).}

\item{keystring}{Character string met kolomnamen gescheiden door punt (bijv. "a.b.c").}

\item{new_col}{Naam van de toe te voegen kolom als string.}

\item{sep}{Scheidingsteken voor samenvoegen van kolommen (default is ".").}

\item{verbose}{Logical, als TRUE wordt een bericht weergegeven na toevoegen.}
}
\value{
NULL (invisible), aanpassing gebeurt by-reference in \code{dt}.
}
\description{
Deze functie voegt aan een data.table een nieuwe kolom toe door opgegeven kolommen 
te concatenaten met een scheidingsteken, alles by-reference.
}
