% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{gen_vector_met_geleidelijk_grotere_stap}
\alias{gen_vector_met_geleidelijk_grotere_stap}
\title{Genereer vector met geleidelijk toenemende stapgroottes}
\usage{
gen_vector_met_geleidelijk_grotere_stap(
  from = 1,
  to = 1e+06,
  steps = c(1, 5, 10, 50, 100, 500, 1000, 5000, 10000, 50000, 1e+05, 5e+05),
  approx_total_length = 500,
  min_per_step = 1
)
}
\arguments{
\item{from}{Numerieke ondergrens van de vector (default = 1)}

\item{to}{Numerieke bovengrens van de vector (default = 1e6)}

\item{steps}{Vector met mogelijke stapgroottes (standaard: \code{c(1, 5, 10, 50, 100, 500, 1000, 5000, 10000, 50000, 100000, 500000)})}

\item{approx_total_length}{Gewenste (ongeveer) totale lengte van de outputvector (default = 500)}

\item{min_per_step}{Minimum aantal punten per stapgrootte (default = 1)}
}
\value{
Een oplopende vector van unieke waarden tussen `from` en `to`
}
\description{
Genereert een numerieke vector tussen `from` en `to`, waarbij de stapgroottes
geleidelijk toenemen. De stapgroottes zijn vooraf gespecificeerd, en het totaal
aantal elementen benadert `approx_total_length`. De overgang tussen stapgroottes
verloopt geleidelijk om abrupte sprongen te vermijden.
}
\examples{
gen_vector_met_geleidelijk_grotere_stap(300, 100000, approx_total_length = 20)
gen_vector_met_geleidelijk_grotere_stap(1, 1e6, approx_total_length = 100)
}
