% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{genereer_functielijst_per_script_mooi}
\alias{genereer_functielijst_per_script_mooi}
\title{Genereer functielijst per script}
\usage{
genereer_functielijst_per_script_mooi(
  mapnaam = "R",
  output_bestand = "functielijst_per_script.txt",
  sorteer = TRUE,
  naar_dev = FALSE
)
}
\description{
Genereer functielijst per script
}
