% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_integer.R
\name{herstel_meerdere_integerkolommen}
\alias{herstel_meerdere_integerkolommen}
\title{Herstel integergecodeerde kolommen naar oorspronkelijke stringwaarden}
\usage{
herstel_meerdere_integerkolommen(dt, kolommen, map_dt)
}
\arguments{
\item{dt}{Een `data.table` met integergecodeerde kolommen}

\item{kolommen}{Vector van kolomnamen in `dt` die hersteld moeten worden}

\item{map_dt}{Mapping `data.table` met kolommen `orig` en `int`}
}
\value{
Geen returnwaarde; `dt` wordt by reference aangepast
}
\description{
Deze functie converteert integerwaarden in de opgegeven kolommen van `dt` terug naar strings,
gebruikmakend van `map_dt` zoals gegenereerd door `maak_gemeenschappelijke_mapping()`.
}
