% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{maak_acc_cols_gev_ongev}
\alias{maak_acc_cols_gev_ongev}
\title{Maak acc-kolommen op basis van gewenste of ongewenste waarden}
\usage{
maak_acc_cols_gev_ongev(dt, b_col, gv_list, default_mode, suffix = "acc")
}
\arguments{
\item{dt}{Een data.table waarin kolommen worden toegevoegd.}

\item{b_col}{Character vector van kolomnamen waarvoor acc-kolommen gemaakt worden.}

\item{gv_list}{Named list met per kolom de gewenste of ongewenste waarden.}

\item{default_mode}{Named character vector met per kolom "gev" of "ongev".
Indien mode = "gev": waarden in gv worden behouden, rest = 0.
Indien mode = "ongev": waarden in gv worden 0, rest wordt behouden.
Indien kolom ontbreekt in gv_list, dan:
   - bij mode = "gev": hele kolom behouden
   - bij mode = "ongev": hele kolom 0}

\item{suffix}{Suffix voor de naam van de nieuwe kolommen (default = "acc")}
}
\value{
De originele data.table, met extra kolommen toegevoegd (invisibly).
}
\description{
Maak acc-kolommen op basis van gewenste of ongewenste waarden
}
