% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_integer.R
\name{maak_gemeenschappelijke_mapping}
\alias{maak_gemeenschappelijke_mapping}
\title{Maak een gemeenschappelijke integermapping van stringwaarden}
\usage{
maak_gemeenschappelijke_mapping(dt1, cols1, dt2, cols2)
}
\arguments{
\item{dt1}{Eerste `data.table` met kolommen `cols1`}

\item{cols1}{Vector van kolomnamen in `dt1` (karakter)}

\item{dt2}{Tweede `data.table` met kolommen `cols2`}

\item{cols2}{Vector van kolomnamen in `dt2` (karakter)}
}
\value{
`data.table` met kolommen `orig` (originele waarden) en `int` (bijbehorende integercodes)
}
\description{
Deze functie genereert een `data.table` die een unieke mapping bevat van stringwaarden uit kolommen in `dt1` en `dt2`
naar een geheel getal. Optioneel wordt `NA` ook gemapt.
}
