% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_pipeline_variant.R
\name{matching_pipeline_variant}
\alias{matching_pipeline_variant}
\title{Wrapper voor matching_pipeline variant 1, 2 of 3}
\usage{
matching_pipeline_variant(
  dt1 = NULL,
  dt2,
  sep = ".",
  kol_a = NULL,
  kol_a_cl = NULL,
  bx = character(),
  cx = character(),
  bx_cl = NULL,
  cx_cl = NULL,
  ag_a = NULL,
  kol_N = NULL,
  variant = c("var1", "var2", "var3"),
  marge,
  na.rm = FALSE,
  auto_bw = TRUE,
  timing = FALSE,
  row_limit = 5e+06,
  nomatch = 0L,
  ..verbose = TRUE,
  show_progress = FALSE,
  logfile = NULL
)
}
\arguments{
\item{dt1}{Eerste data.table (alleen nodig bij variant 1 of 2)}

\item{dt2}{Tweede data.table}

\item{sep}{Scheidingsteken voor samengestelde kolommen}

\item{kol_a}{Naam van groepskolom}

\item{kol_a_cl}{Naam van groepskolom_cl (alleen nodig bij variant 3)}

\item{bx}{Vector met kolomnamen voor exacte matching}

\item{cx}{Vector met kolomnamen voor intervalschatting}

\item{bx_cl}{Kolomnamen van bx_cl (alleen voor variant 3)}

\item{cx_cl}{Kolomnamen van cx_cl (alleen voor variant 3)}

\item{ag_a}{Aggregatietabel met groepsgroottes (alleen voor variant 3)}

\item{kol_N}{Naam van kolom met groepsgrootte in ag_a}

\item{variant}{Te kiezen variant: "var1" (U_exact), "var2" (s_exact), "var3" (s_geschat)}

\item{marge}{Margeparameter voor intervalgrenzen}

\item{na.rm}{Of NA’s moeten worden verwijderd bij samenvoegen}

\item{.verbose}{Of voortgangsmeldingen getoond moeten worden}
}
\value{
data.table met toegevoegde kolommen, afhankelijk van variant
}
\description{
Roept de juiste variant aan en loopt over alle kolommen in `bx` en `cx`.
}
