% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_project_and_original_names.R
\name{rename_to_original_names}
\alias{rename_to_original_names}
\title{Hernoem kolomnamen van projectnaam naar originele naam}
\usage{
rename_to_original_names(vars, naammap)
}
\arguments{
\item{vars}{Character vector van kolomnamen}

\item{naammap}{Named character vector, waarbij `names(naammap)` originele namen zijn,
en `naammap` de corresponderende projectnamen}
}
\value{
Character vector met hernoemde kolommen
}
\description{
Hernoem kolomnamen van projectnaam naar originele naam
}
\examples{
naammap <- c(LFT = "leeftijd", WT = "gewicht")
rename_to_original_names(c("LFT", "WT"), naammap)

}
