% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{rename_with_prefix}
\alias{rename_with_prefix}
\title{Hernoemt kolommen met een prefix}
\usage{
rename_with_prefix(dt, prefix)
}
\arguments{
\item{dt}{Een data.table met kolommen om te hernoemen.}

\item{prefix}{Een prefix toe te voegen aan alle kolomnamen.}
}
\value{
De aangepaste data.table.
}
\description{
Hernoemt kolommen met een prefix
}
\examples{
dt <- data.table::data.table(x = 1:3)
rename_with_prefix(dt, "test")

}
