% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_var12.R
\name{tel_b1_in_interval_rond_b2_foverlaps}
\alias{tel_b1_in_interval_rond_b2_foverlaps}
\title{Functie 4 bereken exacte matches Cx var2}
\usage{
tel_b1_in_interval_rond_b2_foverlaps(
  dt1,
  dt2,
  kol_a,
  kol_c,
  marge,
  row_limit,
  nomatch,
  ...verbose,
  show_progress,
  logfile
)
}
\description{
marge-interval rond b2 in dt2, matchend met b1 in dt1
}
