% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_var12.R
\name{tel_b2_in_interval_rond_b1_foverlaps}
\alias{tel_b2_in_interval_rond_b1_foverlaps}
\title{Functie 2 bereken exacte matches Cx var1}
\usage{
tel_b2_in_interval_rond_b1_foverlaps(
  dt1,
  dt2,
  kol_a,
  kol_c,
  marge,
  row_limit,
  nomatch,
  ...verbose,
  show_progress,
  logfile
)
}
\description{
Functie 2 bereken exacte matches Cx var1
}
\details{
marge rond b1 in dt1, matchend met b2 in dt2
}
