% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_var12.R
\name{tel_exact_match_per_rij_dt1}
\alias{tel_exact_match_per_rij_dt1}
\title{Functie 1 bereken exacte matches Bx var1}
\usage{
tel_exact_match_per_rij_dt1(dt1, dt2, kol_a, kol_b)
}
\description{
Functie 1 bereken exacte matches Bx var1
}
\details{
exacte matches uit dt2 toevoegen aan dt1 (schaalbaar)
}
