% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bx_var12.R
\name{tel_exact_match_per_rij_dt2}
\alias{tel_exact_match_per_rij_dt2}
\title{Functie 3 bereken exacte matches Bx var2}
\usage{
tel_exact_match_per_rij_dt2(dt1, dt2, kol_a, kol_b)
}
\description{
Functie 3 bereken exacte matches Bx var2
}
\details{
exacte matches uit dt1 toevoegen aan dt2 (schaalbaar)
}
