% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vervang_kolommen_als_dev_per_groep.R
\name{vervang_kolommen_als_dev_per_groep}
\alias{vervang_kolommen_als_dev_per_groep}
\title{Vervang kolomwaarden per groep op basis van verdeling binnen groep}
\usage{
vervang_kolommen_als_dev_per_groep(
  dt,
  kol_a,
  cols,
  p_dev,
  p_frac,
  suffix = "_x"
)
}
\arguments{
\item{dt}{Een data.table}

\item{kol_a}{Groepskolomnaam (string)}

\item{cols}{Vector van kolomnamen die aangepast worden}

\item{p_dev}{Kans om een waarde te vervangen (0–1)}

\item{p_frac}{Kans om te vervangen door NA in plaats van bestaande waarde (0–1)}

\item{suffix}{Suffix voor de nieuwe kolommen (standaard "_x")}
}
\description{
Vervang kolomwaarden per groep op basis van verdeling binnen groep
}
