% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_integer.R
\name{vervang_meerdere_stringkolommen}
\alias{vervang_meerdere_stringkolommen}
\title{Vervang meerdere stringkolommen in twee `data.tables` met gemeenschappelijke integercodes}
\usage{
vervang_meerdere_stringkolommen(dt1, cols1, dt2, cols2)
}
\arguments{
\item{dt1}{Eerste `data.table`}

\item{cols1}{Vector van kolomnamen in `dt1` om te vervangen}

\item{dt2}{Tweede `data.table`}

\item{cols2}{Vector van kolomnamen in `dt2` om te vervangen}
}
\value{
Een `data.table` met kolommen `orig` en `int` (de gebruikte mapping)
}
\description{
Genereert een mapping met `maak_gemeenschappelijke_mapping()` en vervangt stringwaarden in `dt1` en `dt2`
door integers.
}
