% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_integer.R
\name{vervang_stringkolommen_met_bestaande_mapping}
\alias{vervang_stringkolommen_met_bestaande_mapping}
\title{Vervang stringwaarden in kolommen van een `data.table` op basis van bestaande mapping}
\usage{
vervang_stringkolommen_met_bestaande_mapping(dt1, cols1, map_dt)
}
\arguments{
\item{dt1}{Een `data.table` waarvan kolommen vervangen worden}

\item{cols1}{Vector van kolomnamen in `dt1` die vervangen worden}

\item{map_dt}{Mapping `data.table` met kolommen `orig` en `int`}
}
\value{
Geen returnwaarde; `dt1` wordt by reference aangepast
}
\description{
Vervang stringwaarden in kolommen van een `data.table` op basis van bestaande mapping
}
