% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voeg_samengestelde_kolom_toe.R
\name{voeg_samengestelde_kolom_toe}
\alias{voeg_samengestelde_kolom_toe}
\title{Voeg een samengestelde characterkolom toe aan een data.table}
\usage{
voeg_samengestelde_kolom_toe(
  dt,
  v,
  kolomnaam = "nieuwe_kolom",
  sep = ".",
  na_handling = "keep",
  na_vervanger = "<leeg>"
)
}
\arguments{
\item{dt}{Een data.table waarin de kolom moet worden toegevoegd.}

\item{v}{Character vector van kolomnamen om samen te voegen. Mag NULL of leeg zijn.}

\item{kolomnaam}{Naam van de toe te voegen kolom. Standaard `"nieuwe_kolom"`.}

\item{sep}{Teken waarmee kolomwaarden worden samengevoegd (alleen bij lengte > 1). Standaard `"."`.}

\item{na_handling}{Hoe om te gaan met NA's. Mogelijke waarden: `"keep"` (standaard), `"remove"`, `"as_string"`, `"replace"`.}

\item{na_vervanger}{String waarmee NA's worden vervangen indien `na_handling = "replace"`.}
}
\value{
De data.table wordt aangepast in-place (onzichtbaar geretourneerd).
}
\description{
Voegt een kolom toe aan een data.table op basis van opgegeven kolommen. Als geen kolommen zijn opgegeven (NULL of lengte 0),
wordt de kolom gevuld met `"All"`. Als er meerdere kolommen worden opgegeven, worden de waarden samengevoegd met een separator.
NA-afhandeling kan worden aangepast via de `na_handling`-parameter.
}
\examples{
library(data.table)
dt <- data.table::data.table(A = c("x", NA, "z"), B = c(1, 2, NA))

voeg_samengestelde_kolom_toe(dt, c("A", "B"), "samen")
voeg_samengestelde_kolom_toe(dt, NULL, "alle_rijen")
voeg_samengestelde_kolom_toe(dt, "A", "alleen_A")
voeg_samengestelde_kolom_toe(dt, c("A", "B"), "verwijder_NA", na_handling = "remove")
voeg_samengestelde_kolom_toe(dt, c("A", "B"), "vervang_NA", na_handling = "replace", na_vervanger = "leeg")

}
