% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var3_cx.R
\name{estimate_c1_counts_in_c2_intervals}
\alias{estimate_c1_counts_in_c2_intervals}
\title{Functie schatting Cx var3}
\usage{
estimate_c1_counts_in_c2_intervals(
  dt2,
  kol_a,
  kol_c,
  kol_a_cl,
  kol_c_cl,
  ag_a,
  kol_N,
  m,
  auto_bw = TRUE,
  row_limit,
  nomatch = 0L,
  ...verbose,
  show_progress = F,
  logfile,
  timing = F
)
}
\description{
Functie schatting Cx var3
}
\details{
schatting Cx var3
}
