% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_pipeline_variant.R
\name{matching_pipeline_variant}
\alias{matching_pipeline_variant}
\title{matching_pipeline_variant}
\usage{
matching_pipeline_variant(
  dt1 = NULL,
  dt2,
  sep = ".",
  kol_a = NULL,
  kol_a_cl = NULL,
  bx = character(),
  cx = character(),
  bx_cl = NULL,
  cx_cl = NULL,
  ag_a = NULL,
  kol_N = NULL,
  variant = c("var1", "var2", "var3"),
  marge,
  na.rm = FALSE,
  auto_bw = TRUE,
  timing = FALSE,
  row_limit = 5e+06,
  nomatch = 0L,
  ..verbose = TRUE,
  show_progress = FALSE,
  logfile = NULL,
  grid_mode
)
}
\description{
matching_pipeline_variant
}
\details{
matching_pipeline_variant
}
